+++ .flutter-plugins
@@ -1,4 +1,5 @@
 # This is a generated file; do not edit or check into version control.
+device_info_plus=/Users/divyanshgandhi/Developer/flutter/.pub-cache/hosted/pub.dartlang.org/device_info_plus-8.0.0/
 flutter_plugin_android_lifecycle=/Users/divyanshgandhi/Developer/flutter/.pub-cache/hosted/pub.dartlang.org/flutter_plugin_android_lifecycle-2.0.7/
 image_picker=/Users/divyanshgandhi/Developer/flutter/.pub-cache/hosted/pub.dartlang.org/image_picker-0.8.6+1/
 image_picker_android=/Users/divyanshgandhi/Developer/flutter/.pub-cache/hosted/pub.dartlang.org/image_picker_android-0.8.5+5/
