+++ example/.gitignore
@@ -43,9 +43,6 @@ app.*.map.json
 /android/app/profile
 /android/app/release
 
-
-android/
-ios/
 linux/
 web/
 macos/
