+++ example/android/app/src/main/kotlin/com/example/example/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.example.example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}
