+++ example/ios/Flutter/Generated.xcconfig
@@ -4,8 +4,8 @@ FLUTTER_APPLICATION_PATH=/Users/divyanshgandhi/Work/LikeMinds-Flutter-Feed-SX/ex
 COCOAPODS_PARALLEL_CODE_SIGN=true
 FLUTTER_TARGET=/Users/divyanshgandhi/Work/LikeMinds-Flutter-Feed-SX/example/lib/main.dart
 FLUTTER_BUILD_DIR=build
-FLUTTER_BUILD_NAME=1.0.0
-FLUTTER_BUILD_NUMBER=1
+FLUTTER_BUILD_NAME=0.5.0
+FLUTTER_BUILD_NUMBER=0.5.0
 EXCLUDED_ARCHS[sdk=iphonesimulator*]=i386
 EXCLUDED_ARCHS[sdk=iphoneos*]=armv7
 DART_DEFINES=RkxVVFRFUl9XRUJfQVVUT19ERVRFQ1Q9dHJ1ZQ==
