+++ example/ios/Flutter/flutter_export_environment.sh
@@ -5,8 +5,8 @@ export "FLUTTER_APPLICATION_PATH=/Users/divyanshgandhi/Work/LikeMinds-Flutter-Fe
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
 export "FLUTTER_TARGET=/Users/divyanshgandhi/Work/LikeMinds-Flutter-Feed-SX/example/lib/main.dart"
 export "FLUTTER_BUILD_DIR=build"
-export "FLUTTER_BUILD_NAME=1.0.0"
-export "FLUTTER_BUILD_NUMBER=1"
+export "FLUTTER_BUILD_NAME=0.5.0"
+export "FLUTTER_BUILD_NUMBER=0.5.0"
 export "DART_DEFINES=RkxVVFRFUl9XRUJfQVVUT19ERVRFQ1Q9dHJ1ZQ=="
 export "DART_OBFUSCATION=false"
 export "TRACK_WIDGET_CREATION=true"
