+++ lib/credentials.dart
@@ -3,10 +3,10 @@
 /// You can use the default credentials provided by the Flutter Sample community.
 
 //These are Koshiqa Beta Sample community credentials
-const String BETA_API_KEY = "7359363e-038b-4b8d-ba00-1f6d4e0266f7";
-const String BETA_BOT_ID = "3a4c74fa-d74d-4b1c-8c0e-c8cbdbc4cc70";
-const String BETA_TEST_USER = "";
+// const String BETA_API_KEY = "7359363e-038b-4b8d-ba00-1f6d4e0266f7";
+// const String BETA_BOT_ID = "3a4c74fa-d74d-4b1c-8c0e-c8cbdbc4cc70";
+// const String BETA_TEST_USER = "";
 
 //These are Flutter Sample community credentials
-// const String BETA_API_KEY = "bad53fff-c85a-4098-b011-ac36703cc98b";
-// const String BETA_BOT_ID = "22b6a64f-66bf-4bca-800e-b40ca66f924d";
+const String BETA_API_KEY = "bad53fff-c85a-4098-b011-ac36703cc98b";
+const String BETA_BOT_ID = "22b6a64f-66bf-4bca-800e-b40ca66f924d";
