+++ pubspec.yaml
@@ -27,11 +27,15 @@ dependencies:
   url_launcher: ^6.1.7
   video_player: ^2.4.10
   timeago: 
-  likeminds_feed: 0.5.0
+  # likeminds_feed: 0.5.0
   image_picker: ^0.8.6+1
   collection: ^1.16.0
   lazy_load_scrollview: ^1.3.0
   infinite_scroll_pagination: ^3.2.0
+  device_info_plus: ^8.0.0
+
+  likeminds_feed: 
+    path: ../LikeMinds-Flutter-SDK
   
 
 dev_dependencies:
